make sequential producer reducing the load in mitre_wide_attack.py
